- Use jwt token per user, encoding the mattermost display name - Use "anonymous"-token in mobile app, as custom message types are not supported - render token validity time on client side to get user timezone right - change client for http requests in frontend to redux client - remove sudo in dockerfile as it poses security risk: script can be run with sudo if desired and user running the script should be added to docker group anyways - make code compile and pass style checks on a clean installation (e.g. docker image) and pull in some missing node build dependencies
User specific tokens are requested, once the message renders. A new endpoint is exposed at /api/v1/meetings/gentoken to request tokens. Tokens are regenerated on every new reder fo the site! The tokens validity time thus does not double as a limit for the time a user has to join the meeting anymore, but the actual validity time of a generated token! For mobile users the old behavior applies.
Security considderations: It would in theory be possible to request token generation for arbitrary users and channels, when forging a request against the endpoint including a full, valid "mattermost-header" and knowing a valid userID and channelID that user is part of.